+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 ------------------
 
 * Fix login password field
+* Enable sourcemaps generation when using --debug param on Gulp
 
 1.0.6 (2015-11-19)
 ------------------